Use nanosleep instead of usleep
This change converts from usleep to nanosleep. This handles EINTR
more naturally, by returning a remainder, and is also more granular
than usleep().

+++ utils/s2n_random.c
@@ -113,9 +113,9 @@ int s2n_get_urandom_data(struct s2n_blob *blob)
     return 0;
 }
 
-int s2n_public_random(int max)
+int64_t s2n_public_random(int64_t max)
 {
-    unsigned int r;
+    uint64_t r;
 
     gt_check(max, 0);
 
@@ -136,7 +136,7 @@ int s2n_public_random(int max)
          * But since 'max' is an int and INT_MAX is <= UINT_MAX / 2,
          * in the worst case we discard 25% - 1 r's.
          */
-        if (r < (UINT_MAX - (UINT_MAX % max))) {
+        if (r < (UINT64_MAX - (UINT64_MAX % max))) {
             return r % max;
         }
     }
